+++ app/Http/Controllers/ServiceController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Service;
 use Illuminate\Http\Request;
 use App\Http\Requests\ServiceRequest;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class ServiceController extends Controller
@@ -21,11 +22,17 @@ class ServiceController extends Controller
 
     public function create()
     {
+        if (Auth::user()->role === 'admin_branch') {
+            return redirect()->route('services.index')->with('error', 'Bạn không có quyền truy cập.');
+        }
         return view('admin.services.create');
     }
 
     public function store(ServiceRequest $request)
     {
+        if (Auth::user()->role === 'admin_branch') {
+            return redirect()->route('services.index')->with('error', 'Bạn không có quyền truy cập.');
+        }
         $data = $request->validated();
 
         // Nếu có ảnh thì lưu vào storage/app/public/services
@@ -46,12 +53,20 @@ class ServiceController extends Controller
 
     public function edit(Service $service)
     {
+
+        if (Auth::user()->role === 'admin_branch') {
+            return redirect()->route('services.index')->with('error', 'Bạn không có quyền chỉnh sửa.');
+        }
         return view('admin.services.edit', compact('service'));
     }
 
     public function update(ServiceRequest $request, Service $service)
     {
 
+
+        if (Auth::user()->role === 'admin_branch') {
+            return redirect()->route('services.index')->with('error', 'Bạn không có quyền truy cập.');
+        }
         $data = $request->validated();
 
         // Nếu có ảnh mới được upload
@@ -66,7 +81,7 @@ class ServiceController extends Controller
         }
 
         $service->update($data);
-        
+
         // Lấy số trang từ request
         $currentPage = $request->input('page', 1);
 
@@ -76,6 +91,10 @@ class ServiceController extends Controller
 
     public function destroy(Service $service)
     {
+
+        if (Auth::user()->role === 'admin_branch') {
+            return redirect()->route('services.index')->with('error', 'Bạn không có quyền xóa dịch vụ.');
+        }
         // Xóa ảnh nếu tồn tại
         if ($service->image && Storage::disk('public')->exists($service->image)) {
             Storage::disk('public')->delete($service->image);
